deleted unneed test files, moved code from question.test.js to questions.test.js
