--- tests/timer.test.js
@@ -1,9 +1,0 @@
-// var startMyTimer = require("./timer");
-
-// describe("timer function testing", () =>{
-//     test("timer is returned", () =>{
-//         expect(startMyTimer(){
-//             var clock = 
-//         })
-//     })
-// }
